update/ adjust WrapperFile components
